Milestone 7 Finishing Touches
The REST API under /service is separate from the web application /, /library, /login, etc. so we don't redirect everything to /service
The Rest Controller needed mapping and the functions need to return response entities and HTTP statuses
The pathing is case sensitive /service != /Service
We ended up not needing DataAccess, please pull from main
The request matcher in the security chain needed to add /service, not replace /library
We redirect everything on the web application to the web application. Successful login routes to /library, not /service/library
We replaced SessionState with SessionLibrary. Please pull from main